SDOCM00053743: issues when setting rotation if rotation is not enabled fixed
If rotation is not enabled through bootargs and if application
tries to set the rotation angle, then FBDEV drivers was accepting
the rotation value and was changing the xres and yres. There
was no return path from this.

Now driver checks for the rotation_type, and if changes/configures
rotation angle only if it is enabled.