PM : cpuidle - Update statistics for correct state
When 'enable_off_mode' is 0, and (mpu_state < PWRDM_POWER_RET)
the local variables mpu_state and core_state are modified; but
the usage count for the original state selected by the governor
are updated.

This patch updates the 'last_state' in the cpuidle driver to ensure
that statistics for the correct state are updated.

When the OFF mode is disabled, all the states where MPU can go to
OFF are disabled.
(cherry picked from commit cc3518c8a66524505db1ff4fd3a8ab4488c1a74f)