Audio: Rebasing misc kernel files from l-o tip
Since the ASoC framework has changed a lot with respect to the
current kernel version, this commit pulls the desired changes
in few kernel source/header files, which are required to compile
the audio component. This update is done to take the twl4030
audio codec patches from the community, which solve the beep
noise issue.