Introduce default omap3_evm_defconfig for PSP release
The community tree no longer maintains board specific
default configurations. However, it is required to
maintain ease of use for developers and customers.

This patch adds the default configuration for omap3evm
to be used in the PSP codebase.

The contents of this configuration were derived from the
PSP 03.00.02.07 release.