OMAP3: PM: Set/reset T2 bit for Smartreflex on TWL.
The smartreflex bit on twl4030 needs to be enabled by default irrespective of
whether smartreflex module is enabled on the OMAP side or not. This is because
without this bit enabled the voltage scaling through vp forceupdate does not
function properly on OMAP3.API added 'omap3_twl_set_sr_bit' with parameter to
set/clear SR bit. It is cleared for platforms where voltage is not scaled using
vpforceupdate or vc_bypass Method. In those cases 'omap3_twl_set_sr_bit' is
called from board file, to make sure this bit is not overwritten in 'omap3_twl_init',
a flag 'twl_sr_enable' is added.

The call to omap3_twl_set_sr_bit() has also been added to omap3_twl_init(),
the twl4030_vsel_to_uv() (where it was originally added) is no longer getting
called (except when using debugfs).

Signed-off-by: Thara Gopinath <thara@ti.com>
Signed-off-by: Shweta Gulati <shweta.gulati@ti.com>
Signed-off-by: Abhilash Vadakkepat Koyamangalath <abhilash.kv@ti.com>