AM35x: Workaround to continue using generic OMAP3 hwmods
This patch implements workaround to reuse the hwmods
defined for OMAP3. Else, the master-slave dependency
cascades into duplicating all of the hwmods.

Without this workaround, the AM3517 EVM fails to boot
due to kernel panic while initializing the SR hwmods.

Until real solution is implemented, this workaround
can be considered.

Signed-off-by: Sanjeev Premi <premi@ti.com>
Signed-off-by: Amit Pundir <amitpundir@ti.com>