ALSA SoC: TWL6040: add switch based headset detection support
Android relies on a switch class driver to report uEvent
for headset status. This adds support for basic headset status
reporting to the twl6040 codec driver. The switch is registered
only if the codec private data indicates it should be used.
This allows boards that don't use jack detection mechanism to
register the switch for headset detection elsewhere.

Change-Id: Ie7762f9a99246b30fb4f1ee4a9e63e62079d334e
Signed-off-by: Chris Kelly <c-kelly@ti.com>