Temporarily fix display flashing issue when enabling off mode
Don't allow PER power domain to go to OFF in idle state transitions.
Restrict PER to RETENTION if it was supposed to go to OFF
This is a temporary fix for display flashing issue which occurs
when off mode is enabled. In suspend PER will go to OFF mode if
enable_off_mode was set.

Signed-off-by: Arun Joseph <arunjoseph@ti.com>