omap2+: serial: Fix random hang when sleep is enabled
With a minimal kernel image, the linux kernel would hang
randomly - usually without any dump.

Only on rare occasions, this dump could be observed:
  Unhandled fault: external abort on non-linefetch (0x1028) at 0xfa3070a0
  Internal error: : 1028 [#1]
  last sysfs file: /sys/power/state
  Modules linked in:
  CPU: 0    Not tainted  (2.6.37-14305-g3767dd6-dirty #155)
  PC is at omap_uart_allow_sleep+0x40/0xe0
  LR is at omap_uart_can_sleep+0x38/0x64
  pc : [<c00483dc>]    lr : [<c00484b8>]    psr: 200000d3
  sp : c0429fa8  ip : c0438880  fp : 00000000
  r10: 0000001f  r9 : 413fc082  r8 : 80028110
  r7 : c042cd0c  r6 : c00299dc  r5 : 00000000  r4 : cf8052c0
  r3 : 00000001  r2 : fa3070a0  r1 : 00040000  r0 : cf8052c0
  Flags: nzCv  IRQs off  FIQs off  Mode SVC_32  ISA ARM  Segment kernel

The script to recreate the hang/crash is:
  mount -t debugfs debugfs /mnt
  echo 5 > /sys/devices/platform/omap/omap_uart.0/sleep_timeout
  echo 5 > /sys/devices/platform/omap/omap_uart.1/sleep_timeout
  echo 5 > /sys/devices/platform/omap/omap_uart.2/sleep_timeout
  echo 5 > /sys/devices/platform/omap/omap_uart.3/sleep_timeout

  echo 1 > /mnt/pm_debug/sleep_while_idle
  echo 1 > /mnt/pm_debug/enable_off_mode

The problem occurs when accessing the registers UART registers
are accessed when the clocks are disabled.

This patch moves steps to enable/ disable wakeup into function
omap_uart_prepare_idle() before the clocka are disabled. This
makes the usage similar to omap_uart_resume_idle().

With register access happening only when the clocks are enabled,
there is no possibility of hang/ crash - as originally observed.

Signed-off-by: Sanjeev Premi <premi@ti.com>